requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.7.0

# LangGraph and LangChain
langgraph>=0.2.60
langchain>=0.3.12
langchain-core>=0.3.28

# LLM SDKs
google-genai>=1.0.0
anthropic>=0.40.0
openai>=1.58.0

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
redis==5.2.1

# Vector DB
qdrant-client>=1.12.0

# Embeddings
sentence-transformers>=3.3.0

# Web scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
duckduckgo-search>=5.0.0

# PDF parsing
pypdf==5.1.0
pdfplumber==0.11.4

# Data processing
pandas==2.2.3
numpy>=1.26.4

# Web UI
streamlit>=1.41.0

# API (Phase 2)
fastapi>=0.115.0
uvicorn>=0.34.0

# Utilities
python-dateutil==2.9.0
loguru==0.7.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
